requirements/base.txt
aiohttp==3.8.4
aiohttp-proxy==0.1.2
aiosignal==1.3.1
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
loguru==0.7.0
mccabe==0.7.0
multidict==6.0.4
names==0.3.0
openpyxl==3.1.2
peewee==3.16.2
platformdirs==3.4.0
progress==1.6
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
python-socks==2.2.0
requests==2.28.2
rsa==4.9
Telethon==1.28.2
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.9.2
